Add two Frames
