+++ src/test/java/school/redrover/FreestyleProjectTest.java
@@ -199,7 +199,7 @@ public class FreestyleProjectTest extends BaseTest {
     public void testCheckFreestyleProjectViaBreadcrumb() {
         List<String> itemListInsideFolder = new HomePage(getDriver())
                 .openDashboardBreadcrumbsDropdown()
-                .clickMyViewsFromDropdown()
+                .getHeader().clickMyViewsFromDropdown()
                 .clickBreadcrumbAll()
                 .clickJobNameBreadcrumb(FOLDER_NAME)
                 .getItemListInsideFolder();
