+++ src/test/java/school/redrover/PipelineTest.java
@@ -85,7 +85,7 @@ public class PipelineTest extends BaseTest {
                 .setItemName(PIPELINE_NAME)
                 .selectPipelineAndClickOk()
                 .clickLogo()
-                .searchProjectByName(PIPELINE_NAME, new PipelineProjectPage(getDriver()))
+                .getHeader().searchProjectByName(PIPELINE_NAME, new PipelineProjectPage(getDriver()))
                 .getProjectName();
 
         Assert.assertEquals(actualPipelineName, PIPELINE_NAME);
