+++ src/test/java/school/redrover/model/HomePage.java
@@ -94,18 +94,6 @@ public class HomePage extends BaseSideMenuPage<HomePage> {
     @FindBy(xpath = "//a[contains(@href, '/move')]")
     private WebElement moveOption;
 
-    @FindBy(css = "[class$='am-button security-am']")
-    private WebElement warningIcon;
-
-    @FindBy(xpath = "//div[@role='alert']")
-    private WebElement warningTooltipLocator;
-
-    @FindBy(xpath = "//a[contains(text(),'Manage Jenkins')]")
-    private WebElement manageJenkinsTooltipLink;
-
-    @FindBy(xpath = "//button[@name='configure']")
-    private WebElement configureTooltipButton;
-
     @FindBy(xpath = "//a[@href='/asynchPeople/']")
     private WebElement peopleButton;
 
@@ -237,13 +225,6 @@ public class HomePage extends BaseSideMenuPage<HomePage> {
         freestyleItem.click();
 
         return new FreestyleProjectPage(getDriver());
-    }
-
-    public ViewAllPage clickMyViewsFromDropdown() {
-        openHeaderUsernameDropdown();
-        getDriver().findElement(By.cssSelector("[href$='admin/my-views']")).click();
-
-        return new ViewAllPage(getDriver());
     }
 
     public ManageJenkinsPage clickManageJenkins() {
@@ -470,26 +451,6 @@ public class HomePage extends BaseSideMenuPage<HomePage> {
     public MovePage chooseFolderToMove() {
         getWait5().until(ExpectedConditions.visibilityOf(moveOption)).click();
         return new MovePage(getDriver());
-    }
-
-    public HomePage clickWarningIcon() {
-        warningIcon.click();
-        return this;
-    }
-
-    public String getWarningTooltipText() {
-        WebElement warningTooltipText = getWait5().until(ExpectedConditions.visibilityOf(warningTooltipLocator));
-        return warningTooltipText.getText();
-    }
-
-    public ManageJenkinsPage clickManageJenkinsTooltipLink() {
-        getWait5().until(ExpectedConditions.elementToBeClickable(manageJenkinsTooltipLink)).click();
-        return new ManageJenkinsPage(getDriver());
-    }
-
-    public SecurityPage clickConfigureTooltipButton() {
-        getWait5().until(ExpectedConditions.elementToBeClickable(configureTooltipButton)).click();
-        return new SecurityPage(getDriver());
     }
 
     public PeoplePage clickPeopleOnSidebar() {
